some findbugs fixes
svn path=/trunk/; revision=46903


Former-commit-id: 40602bb09fda3daefcd4f32b76ab4023ddf1f72a

+++ v3/webservices/jsr109-impl/src/main/java/org/glassfish/webservices/monitoring/WebServiceTesterServlet.java
@@ -716,13 +716,15 @@ public class WebServiceTesterServlet extends HttpServlet implements MessageListe
     
     private void deleteDir(File path) {
 
-        File[] files = path.listFiles();
-        for (File f : files) {
-            if (f.isDirectory()) {
-                deleteDir(f);
+        if (path.exists() && path.isFile()) {
+            File[] files = path.listFiles();
+            for (File f : files) {
+                if (f.isDirectory()) {
+                    deleteDir(f);
+                }
+                assert f.delete();
             }
-            f.delete();
+            assert path.delete();
         }
-        path.delete();
     }    
 }
